style and new pages
