+++ .vscode/settings.json
@@ -1,0 +1,11 @@
+{
+    "spellright.language": [
+        "en"
+    ],
+    "spellright.documentTypes": [
+        "markdown",
+        "latex",
+        "plaintext",
+        "javascriptreact"
+    ]
+}
